Move MockHUDProvider to separate file
